requirements.txt
streamlit
PyPDF2
python-docx
fuzzywuzzy
python-Levenshtein
